pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mangrove-studio"
version = "0.1.0a1"
description = "AI-powered carbon modeling tool"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "click>=8.0",
    "simpleeval>=1.0",
    "jsonschema>=4.0",
]

[project.optional-dependencies]
ai = [
    "anthropic>=0.40",
    "openpyxl>=3.1",
]
server = [
    "fastapi>=0.110",
    "uvicorn>=0.30",
    "sse-starlette>=2.0",
]
mcp = [
    "mcp[cli]>=1.0",
]
dev = [
    "pytest>=8.0",
    "ruff>=0.5",
]

[project.scripts]
mangrove = "mangrove_studio.cli.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/mangrove_studio"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 120
